move schemas to common package Co-authored-by: Zhan Xue zhan.xue@intel.com Co-authored-by: Eero Tamminen eero.t.tamminen@intel.com Co-authored-by: Tuomas Katila tuomas.katila@intel.com
