+++ deploy/cloud/operator/internal/common/schemas.go
@@ -14,25 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package common
 
-import (
-	"fmt"
-	"net/url"
-	"strings"
-)
+type DeploymentStrategy string
 
-func GetHost(someURL string) (string, error) {
-	// Add scheme if not present
-	if !strings.Contains(someURL, "://") {
-		someURL = "dummy://" + someURL
-	}
-	url, err := url.Parse(someURL)
-	if err != nil {
-		return "", err
-	}
-	if url.Host == "" {
-		return "", fmt.Errorf("no host found in URL %q", someURL)
-	}
-	return url.Host, nil
-}
+const (
+	DeploymentStrategyRollingUpdate               DeploymentStrategy = "RollingUpdate"
+	DeploymentStrategyRecreate                    DeploymentStrategy = "Recreate"
+	DeploymentStrategyRampedSlowRollout           DeploymentStrategy = "RampedSlowRollout"
+	DeploymentStrategyBestEffortControlledRollout DeploymentStrategy = "BestEffortControlledRollout"
+)
